week7_codelab2/common/news_recommendation_service_client.py
import jsonrpclib

URL = "http://localhost:5050/"

client = jsonrpclib.ServerProxy(URL)

def getPreferenceForUser(userId):
    preference = client.getPreferenceForUser(userId)
    print("Preference list: %s" % str(preference))
    return preference
    # preference is a list of different class names ['US', 'Sports','...' ] 