week7_codelab2/requirements.txt
jsonrpclib-pelix
numpy
pika
python-dateutil
pymongo
redis
requests
scipy
sklearn
lxml
newspaper3k